Delete exception unused, and remove import unused
