+++ src/main/java/org/tptacs/application/security/JwtUtils.java
@@ -2,9 +2,6 @@ package org.tptacs.application.security;
 
 import java.security.Key;
 import java.util.Date;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 import io.jsonwebtoken.*;
@@ -15,7 +12,6 @@ import org.tptacs.domain.exceptions.AuthorizationException;
 
 @Component
 public class JwtUtils {
-    private static final Logger logger = LoggerFactory.getLogger(JwtUtils.class);
 
     // TODO extraer a una varialbe de entorno
     private String jwtSecret = "0b70d97e54964bc0ab6d7f9446e831de0b70d97e54964bc0ab6d7f9446e831de";
