+++ src/main/java/org/tptacs/domain/entities/Order.java
@@ -1,12 +1,14 @@
 package org.tptacs.domain.entities;
 
 
-import lombok.Getter;
-import org.tptacs.domain.enums.OrderStatus;
-import org.tptacs.domain.exceptions.NotFoundException;
 import java.time.LocalDateTime;
 import java.util.List;
+
+import org.tptacs.domain.enums.OrderStatus;
+import org.tptacs.domain.exceptions.ResourceNotFoundException;
+
 import jakarta.validation.ValidationException;
+import lombok.Getter;
 
 @Getter
 public class Order {
@@ -44,7 +46,7 @@ public class Order {
     public void removeItem(Item item){
         var result = this.items.removeIf(io -> io.getItem().getId().equals(item.getId()));
         this.lastUpdate = LocalDateTime.now();
-        if(!result) throw new NotFoundException(item.getId(), "item");
+        if(!result) throw new ResourceNotFoundException(item.getId(), "item");
     }
 
 }
