+++ src/main/java/org/tptacs/exceptionHandler/ControllerExceptionHandler.java
@@ -1,17 +1,16 @@
 package org.tptacs.exceptionHandler;
 
+import java.util.Set;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-import org.tptacs.domain.exceptions.ResourceNotFoundException;
-import org.tptacs.domain.exceptions.NotFoundException;
 import org.tptacs.domain.exceptions.RegistrationException;
+import org.tptacs.domain.exceptions.ResourceNotFoundException;
 import org.tptacs.domain.exceptions.ValidationException;
-
-import java.util.Set;
 
 @ControllerAdvice
 public class ControllerExceptionHandler extends ResponseEntityExceptionHandler {
@@ -19,12 +18,6 @@ public class ControllerExceptionHandler extends ResponseEntityExceptionHandler {
     public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
         System.out.println("error no manejado: " + ex.getMessage());
         return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ErrorResponse(HttpStatus.INTERNAL_SERVER_ERROR, ex.getMessage()));
-    }
-
-    @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleNotFoundException(NotFoundException ex) {
-        System.out.println("error: " + ex.getMessage());
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponse(HttpStatus.NOT_FOUND, ex.getMessage()));
     }
 
     @ExceptionHandler(ValidationException.class)
