+++ src/main/java/org/tptacs/presentation/controllers/TestController.java
@@ -3,7 +3,6 @@ package org.tptacs.presentation.controllers;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
